progress on changing the blog bug
made it so new posts are created with a unique code and show up on the website without backend edits. Only new issue is that when the page is refreshed, it submits either nothing or the old information as a post, which needs to be changed, maybe by opening a new webpage to create a post?